Split enterprise and community builds (#2406)
* Split enterprise and community builds

* Fix import path

* Move license file

* Add mock modules

* Load correct version

* Update gitignore

* Update gitignore

* Update dev docs

* Update dev docs

* Update dev docs

* Remove version.json
