+++ .gitignore
@@ -36,7 +36,7 @@ Thumbs.db
 
 # Generated.
 bindata_assetfs.go
-git-version.json
+version.json
 
 # Cypress.
 cypress/fixtures
