+++ src/app/dynamic/common.ts
@@ -1,0 +1,5 @@
+import * as version from '../../assets/config/version.json';
+
+export function getEditionDirName() {
+  return version.edition === 'Community Edition' ? 'community' : 'enterprise';
+}
