+++ src/app/dynamic/community/theming/module.ts
@@ -1,0 +1,5 @@
+import {NgModule} from '@angular/core';
+import {RouterModule} from '@angular/router';
+
+@NgModule({imports: [RouterModule.forChild([])]})
+export class ThemingModule {}
