+++ src/app/dynamic/enterprise/LICENSE
