+++ src/app/dynamic/enterprise/theming/module.ts
@@ -1,10 +1,10 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
-import {SharedModule} from '../shared/shared.module';
 import {StylePickerComponent} from './picker/component';
 import {ColorSchemeService} from './services/color-scheme';
 import {ThemeManagerService} from './services/manager';
 import {ThemeService} from './services/theme';
+import {SharedModule} from '../../../shared/shared.module';
 
 const routes: Routes = [
   {path: '', outlet: 'theming', component: StylePickerComponent},
