+++ src/app/settings/user/user-settings.routing.module.ts
@@ -4,6 +4,7 @@ import {RouterModule, Routes} from '@angular/router';
 import {AuthGuard} from '../../core/services';
 
 import {UserSettingsComponent} from './user-settings.component';
+import {getEditionDirName} from '../../dynamic/common';
 
 const routes: Routes = [
   {
@@ -14,7 +15,9 @@ const routes: Routes = [
       {
         path: '',
         loadChildren: () =>
-          import('../../theming/module').then(module => module.ThemingModule),
+          import(`../../dynamic/${getEditionDirName()}/theming/module`).then(
+            module => module.ThemingModule
+          ),
         data: {preload: true},
       },
     ],
