--- src/app/theming/README.md
@@ -1,2 +1,0 @@
-## Important
-This module is only available in an enterprise version of Kubermatic Dashboard.
