+++ src/environments/environment.prod.ts
@@ -6,7 +6,7 @@ export const environment = {
   name: 'prod',
   production: true,
   configUrl: '/config/config.json',
-  gitVersionUrl: '../assets/config/git-version.json',
+  gitVersionUrl: '../assets/config/version.json',
   refreshTimeBase: 1000, // Unit: ms
   restRoot: '/api/v1',
   wsRoot: `${wsProtocol}//${host}/api/v1/ws`,
