+++ tsconfig.json
@@ -10,6 +10,7 @@
     "experimentalDecorators": true,
     "noImplicitReturns": false,
     "noUnusedLocals": true,
+    "resolveJsonModule": true,
     "strict": false,
     "target": "es5",
     "typeRoots": [
